requirements.txt
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.1
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
schedule==1.2.2
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
